Clean account possible now

+++ controls/action/delete_account.php
@@ -4,6 +4,12 @@ require_once __DIR__.'/../../config-app.php';
 include_once(LIBPATH.'/accounts/get_account_admin.php');
 include_once(LIBPATH.'/accounts/delete_account.php');
 
+include_once(LIBPATH.'/participants/get_participants.php');
+include_once(LIBPATH.'/participants/delete_participant.php');
+
+include_once(LIBPATH.'/bills/get_bills.php');
+include_once(LIBPATH.'/bills/delete_bill.php');
+
 include_once(LIBPATH.'/hashid/validate_hashid.php');
 
 
@@ -15,7 +21,9 @@ $warnArray = array(); //warning messages
 $successArray = array(); //success messages
 $redirect_link ="" ;
 
-if(isset($_POST['submit_delete_account']))
+if(isset($_POST['submit_delete_account'])
+|| isset($_POST['submit_remove_all_participants'])
+|| isset($_POST['submit_remove_all_bills']))
 {
 	$ErrorEmptyMessage = array(
 		'p_hashid_account' => 'No acount provided',
@@ -44,20 +52,52 @@ if(isset($_POST['submit_delete_account']))
 		if(empty($account))
 		{	array_push($errArray, $ErrorMessage['p_hashid_account']); }
 	}
-
-	//Delete the account
-	if(empty($errArray))
-	{
-		$success = delete_account($account['id']);	
-		if(!$success)
-		{array_push($errArray, 'Server error: Problem while attempting to delete an account'); 	}
-			else
-			{
-				array_push($successArray, 'Account has been successfully deleted');
-			}
-	}
 }
 
+if(isset($_POST['submit_delete_account'])
+&& empty($errArray))
+{
+	//Delete the account
+	$success = delete_account($account['id']);	
+	if(!$success)
+	{array_push($errArray, 'Server error: Problem while attempting to delete an account'); 	}
+		else
+		{
+			array_push($successArray, 'Account has been successfully deleted');
+		}
+}
+elseif(isset($_POST['submit_remove_all_participants']) 
+&& empty($errArray))
+{
+	$participants = get_participants($account['id']);
+	//Delete the participants
+	foreach($participants as $parti)
+	{
+		$success = delete_participant($account['id'], $parti['id']);	
+		if(!$success)
+		{array_push($errArray, 'Server error: Problem while attempting to delete a participant'); 	}
+		else
+		{
+			array_push($successArray, 'Participant has been successfully deleted');
+		}
+	}
+}
+elseif(isset($_POST['submit_remove_all_bills']) 
+&& empty($errArray))
+{
+	$bills = get_bills($account['id']);
+	//Delete the participants
+	foreach($bills as $bill)
+	{
+		$success = delete_bill($account['id'], $bill['id']);	
+		if(!$success)
+		{array_push($errArray, 'Server error: Problem while attempting to delete a bill'); 	}
+		else
+		{
+			array_push($successArray, 'Bill has been successfully deleted');
+		}
+	}
+}
 		
 if(!(empty($errArray)))
 {
@@ -72,7 +112,7 @@ if(!(empty($successArray)))
 	$_SESSION['success'] = $successArray;
 }
 
-if(empty($errArray) || empty($account))
+if(!isset($account) || empty($account))
 {
 	$redirect_link = BASEURL;
 }
